bugfixes: hexen2 players now have colourmaps again. doesn't crash with rgb player colours. hexen2 colour previews now work (as well as they did in actual hexen2, anyway) fix possible issue with 'bad type - needs fixing' color command now longer omits 0 values when displaying. readded support for all hexen2 palette lookup types. fixed some hexen2 image positions. using premultiplied alpha for pics as an easy fix for hexen2's various ugly halos. the checkerboard texture is back! fix q1 maps that have more than 32k clipplanes. new features: when associated with mdl etc file extensions, now opens up the modelviewer. also potentially changes gamedir (this applies to bsps too). multiplayer setup menu now supports selecting rgb colours (lower colour remains with the 14 team colours). press shift for the normal/hue colour selection instead. attempt to use bgra for texture uploads by default where available. reimplement possible support for 16bit textures. attempt to implement gamepad support for web port. untested, but should be completeish.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4764 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/server/net_preparse.c
@@ -342,7 +342,7 @@ void NPP_QWWriteString(int dest, char *data)
 {
 	NPP_NQWriteString(dest, data);
 }
-void NPP_QWWriteEntity(int dest, short data)
+void NPP_QWWriteEntity(int dest, int data)
 {
 	NPP_NQWriteEntity(dest, data);
 }
@@ -1487,7 +1487,7 @@ void NPP_NQWriteString(int dest, const char *data)	//replacement write func (nq 
 		nullterms--;
 	NPP_NQCheckFlush();
 }
-void NPP_NQWriteEntity(int dest, short data)	//replacement write func (nq to qw)
+void NPP_NQWriteEntity(int dest, int data)	//replacement write func (nq to qw)
 {
 	NPP_NQCheckDest(dest);
 	if (!bufferlen)
@@ -2150,7 +2150,7 @@ void NPP_QWWriteString(int dest, const char *data)	//replacement write func (nq 
 	NPP_QWCheckFlush();
 #endif
 }
-void NPP_QWWriteEntity(int dest, short data)	//replacement write func (nq to qw)
+void NPP_QWWriteEntity(int dest, int data)	//replacement write func (nq to qw)
 {
 	if (data >= 0x8000)
 	{
